requirements.txt
django==3.2.18
djangorestframework
gunicorn
psycopg2-binary
dj-database-url
whitenoise